Extract ImportTypings
Extract part of ImportTypings as ImportTypingsUtil in importer-portable.
